backend/src/product/utilits.py
from datetime import timedelta, datetime
import bcrypt
import jwt
from backend.src.settings import ALGORITHM, SECRET_KEY

EXPIRATION_TIME = timedelta(minutes=30)


async def hashed_password(password):
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(password.encode('utf-8'), salt)
    return hashed.decode('utf-8')


async def verify_password(password, hash_password):
    return bcrypt.checkpw(password.encode('utf-8'), hash_password.encode('utf-8'))


async def generate_jwt_token(data: dict):
    expiration = datetime.utcnow() + EXPIRATION_TIME
    data.update({'exp': expiration})
    token = jwt.encode(data, SECRET_KEY, algorithm=ALGORITHM)
    return token
