backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ultimap"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
psycopg2 = "^2.9.7"
bcrypt = "^4.0.1"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.3.0"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
